+++ nshm_model_graphql_api/schema/__init__.py
@@ -1,0 +1,3 @@
+"""API schema package."""
+
+from .schema_root import schema_root  # noqa: F401